pyproject.toml
[project]
name = "kadaie1"
version = "0.1.0"
description = "FASTA配列の長さ分析ツール"
authors = [{name = "Your Name", email = "your@email.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = ["biopython"]

[project.scripts]
kadaie_1 = "kadaie1.kadaie_1:main"

[build-system]
requires = ["uv","uv-build"]
build-backend = "uv_build"
